Fixed Translate Service for en, gr.
