+++ src/app/app.component.ts
@@ -1,41 +1,38 @@
-import {HttpClient} from '@angular/common/http';
 import {Component} from '@angular/core';
-import {ActivatedRoute, RouterModule, RouterOutlet} from '@angular/router';
-import {TranslateModule, TranslateService} from '@ngx-translate/core';
-import {TranslateHttpLoader} from '@ngx-translate/http-loader';
+import {RouterModule, RouterOutlet} from '@angular/router';
+import {TranslateService} from '@ngx-translate/core';
 import {FooterComponent} from './components/footer/footer.component';
 import {HeaderComponent} from './components/header/header.component';
-
-export function HttpLoaderFactory(httpClient: HttpClient) {
-  return new TranslateHttpLoader(httpClient);
-}
+import {lang, LangService} from './services/lang.service';
 
 @Component({
   selector: 'app-root',
   standalone: true,
   imports: [
     RouterOutlet,
-    TranslateModule,
+    RouterModule,
     HeaderComponent,
     FooterComponent,
-    RouterModule,
   ],
-  providers: [ActivatedRoute],
+  providers: [
+    LangService,
+  ],
   templateUrl: './app.component.html',
   styleUrl: './app.component.scss',
 })
 export class AppComponent {
   title = 'milonas-jewelry';
 
-  constructor(public translate: TranslateService) {
+  constructor(public translate: TranslateService, public langService: LangService) {
     translate.addLangs([
       'en',
-      'gr',
+      'el',
     ]);
-    translate.setDefaultLang('gr');
+    translate.setDefaultLang('el');
 
-    const browserLang: string = !!translate.getBrowserLang() ? <string>translate.getBrowserLang() : 'gr';
-    translate.use(browserLang.match(/en|gr/) ? browserLang : 'gr');
+    langService.updateLang(!!translate.getBrowserLang() ? <lang>translate.getBrowserLang() : 'el');
+
+    console.log('app = ' + langService.lang$.getValue());
   }
 }
