+++ src/app/app.config.ts
@@ -1,15 +1,28 @@
-import {ApplicationConfig} from '@angular/core';
+import {HttpClient, provideHttpClient} from '@angular/common/http';
+import {ApplicationConfig, importProvidersFrom} from '@angular/core';
 import {provideClientHydration} from '@angular/platform-browser';
 import {provideRouter} from '@angular/router';
+import {TranslateLoader, TranslateModule} from '@ngx-translate/core';
+import {TranslateHttpLoader} from '@ngx-translate/http-loader';
 import {routes} from './app.routes';
 import {DeviceService} from './services/device-service.service';
-import {PointerPositionService} from './services/pointer-position.service';
+
+export function HttpLoaderFactory(httpClient: HttpClient) {
+  return new TranslateHttpLoader(httpClient);
+}
 
 export const appConfig: ApplicationConfig = {
   providers: [
     provideRouter(routes),
     provideClientHydration(),
     DeviceService,
-    PointerPositionService,
+    provideHttpClient(),
+    importProvidersFrom(TranslateModule.forRoot({
+      loader: {
+        provide: TranslateLoader,
+        useFactory: HttpLoaderFactory,
+        deps: [HttpClient],
+      },
+    })),
   ],
 };
