--- src/app/services/pointer-position.service.ts
@@ -1,12 +1,0 @@
-import {Injectable} from '@angular/core';
-
-@Injectable({
-  providedIn: 'root',
-})
-export class PointerPositionService {
-
-  constructor() {
-
-  }
-}
-
