+++ src/assets/i18n/el.json
@@ -1,0 +1,11 @@
+{
+  "footer": {
+    "copyright": "milonas ©2023",
+    "termsOfService": "όροι χρήσης",
+    "cookies": "cookies",
+    "lang": {
+      "en": "αγγλικά",
+      "el": "ελληνικά"
+    }
+  }
+}
