+++ src/assets/i18n/en.json
@@ -2,6 +2,10 @@
   "footer": {
     "copyright": "milonas ©2023",
     "termsOfService": "terms of service",
-    "cookies": "cookies"
+    "cookies": "cookies",
+    "lang": {
+      "en": "english",
+      "el": "greek"
+    }
   }
 }
