--- src/assets/i18n/gr.json
@@ -1,7 +1,0 @@
-{
-  "footer": {
-    "copyright": "milonas ©2023",
-    "termsOfService": "όροι χρήσης",
-    "cookies": "cookies"
-  }
-}
