+++ src/main.ts
@@ -1,20 +1,5 @@
-import {HttpClient, provideHttpClient} from '@angular/common/http';
-import {importProvidersFrom} from '@angular/core';
 import {bootstrapApplication} from '@angular/platform-browser';
-import {RouterModule} from '@angular/router';
-import {TranslateLoader, TranslateModule} from '@ngx-translate/core';
-import {AppComponent, HttpLoaderFactory} from './app/app.component';
-import {routes} from './app/app.routes';
+import {AppComponent} from './app/app.component';
+import {appConfig} from './app/app.config';
 
-bootstrapApplication(AppComponent, {
-  providers: [
-    provideHttpClient(),
-    importProvidersFrom(TranslateModule.forRoot({
-      loader: {
-        provide: TranslateLoader,
-        useFactory: HttpLoaderFactory,
-        deps: [HttpClient],
-      },
-    }), RouterModule.forRoot(routes)),
-  ],
-});
+bootstrapApplication(AppComponent, appConfig);
